refactor(offer): split offer_create.go into focused files (#77)
offer_create.go went from ~1800 lines to ~110, holding only the type,
JSON marshalling, constructors, and flag setters. The rest is grouped
by responsibility:

- offer_create_validate.go (333): Validate, Preflight, Preclaim,
  parsePreflightError, badCurrency, hasExpired, checkAcceptAsset,
  accountInDomain.
- offer_create_apply.go (678): Apply, ApplyCreate, applyGuts,
  applyHybridInSandbox, peekOffer, offerDeleteInView,
  adjustOwnerCountInView, parseFee, getOfferSequence, lsfHybrid.
- offer_quality.go (649): tick-size + quality math
  (applyTickSize, getTickSize, roundToTickSize, qualityToRate,
  multiplyByQuality, divideByQuality, ratToIssuedAmount,
  offerDivRound, offerDivRoundStrict, offerMulRound, maxTickSize).
- offer_amount.go (53): tiny amount predicates
  (isLegalNetAmount, isAmountZeroOrNegative, isAmountNegative,
  zeroAmount, subtractAmounts).

Pure code motion; no behavior change. Same set of tests pass and the
same pre-existing TestOffer_TickSize/.../range subtests fail as on
main.
